requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai>=1.0.0
langchain>=0.0.300
python-dotenv>=1.0.0
pydantic>=1.10.0
jinja2>=3.1.0

# Visualization and utility
mermaid-js>=10.0.0 # (for reference, actual rendering is in frontend)

# If you use graphviz for any backend rendering (optional)
graphviz>=0.20.0

# For file handling and static serving
requests>=2.31.0

pytest>=7.0.0
google-api-python-client
langchain-exa
duckduckgo-search
duckpy